feat: sync collection in the background
aimed only to be used with autosync, so the user can leave the app without having to wait the synchronization to be complete.

to avoid hanging, the ideal solution would be to show a progress dialog if the worker is running.

Right now, that isn't possible because of the synchronized way of building most of the app's entry points (DeckPicker, CardBrowser, Reviewer, etc)

Instead the background sync is simply cancelled.

Since syncing is really fast with the new backend, and that background sync is intended to be done only when the user wants to leave the app, I believe that most cases will get unnoticed