feat: add ReviewerMenuSettingsFragment
just a blank screen for now.

Split from the other commits to make the reviewing process easier

Why using an Activity instead of launching it like the other preferences screens?

Because of the toolbar. The new screen for setting the reviewer menu actions will have a toolbar that serves as a preview of how it will look in the reviewer. Differently from the Preferences' toolbar, it isn't collapsible, so I can't simply do the preview there.

Since all of the options that I thought required a big amount of work, I decided to simply use a standalone activity for now and fix it later.

For phone users, there won't be any visible differences. For tablet users, it just won't shown the lateral navigation bar of the preferences page.